+++ plugins/gs/src/components/Router.tsx
@@ -1,20 +1,16 @@
-import { Entity } from '@backstage/catalog-model';
-import { GS_APP_FLAVOR_LABEL } from './getAppNameFromEntity';
 import { useEntity } from '@backstage/plugin-catalog-react';
 import { Route, Routes } from 'react-router-dom';
 import { EntityDeploymentsContent } from './EntityDeploymentsContent';
 import { entityDeploymentDetailsRouteRef } from '../routes';
 import React from 'react';
 import { EmptyState } from '@backstage/core-components';
-
-export const isGSDeploymentsAvailable = (entity: Entity) =>
-  entity.metadata.labels?.[GS_APP_FLAVOR_LABEL] === 'true';
+import { isEntityDeploymentsAvailable } from './utils/entity';
 
 /** @public */
 export const Router = () => {
   const { entity } = useEntity();
 
-  if (!isGSDeploymentsAvailable(entity)) {
+  if (!isEntityDeploymentsAvailable(entity)) {
     return (
       <EmptyState
         title="Deployments content is not available for this entity"
