+++ plugins/gs/src/components/UI/Date/index.tsx
@@ -1,6 +1,6 @@
 import differenceInMinutes from 'date-fns/differenceInMinutes';
 import React, { useEffect, useMemo, useRef, useState } from 'react';
-import { formatDate, getRelativeDate, parseDate } from '../../helpers';
+import { formatDate, getRelativeDate, parseDate } from '../../utils/helpers';
 import { Tooltip, Typography } from '@material-ui/core';
 
 interface IDateProps extends React.ComponentPropsWithoutRef<typeof Typography> {
