+++ plugins/gs/src/components/utils/cachingColorHash.ts
