+++ plugins/gs/src/components/utils/entity.ts
@@ -1,6 +1,6 @@
 import { Entity } from '@backstage/catalog-model';
 
-export const GS_APP_FLAVOR_LABEL = 'giantswarm.io/flavor-app';
+const GS_APP_FLAVOR_LABEL = 'giantswarm.io/flavor-app';
 const GITHUB_PROJECT_SLUG_ANNOTATION = 'github.com/project-slug';
 
 export const getServiceNameFromEntity = (entity: Entity) =>
@@ -8,3 +8,6 @@ export const getServiceNameFromEntity = (entity: Entity) =>
 
 export const getProjectSlugFromEntity = (entity: Entity) =>
   entity.metadata.annotations?.[GITHUB_PROJECT_SLUG_ANNOTATION];
+
+export const isEntityDeploymentsAvailable = (entity: Entity) =>
+  entity.metadata.labels?.[GS_APP_FLAVOR_LABEL] === 'true';
